ml-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
tensorflow==2.16.1
Pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2
pydantic==2.8.0
